test: extract hook
